needler progress
Current issue is the prefab is not being loaded correctly, will need to find another way to get it in properly

+++ Assets/Player/PlayerEvents.cs
@@ -4,10 +4,10 @@ using UnityEngine;
 
 public class PlayerEvents : MonoBehaviour
 {
-    public delegate void PlayerDamageInstance(GameObject obj, float dmg, Vector3 loc);
+    public delegate void PlayerDamageInstance(GameObject obj, int dmg, Vector3 loc);
     public static event PlayerDamageInstance OnPlayerDamageInstance;
 
-    public static void CallPlayerDamageEvent(GameObject obj, float dmg, Vector3 loc)
+    public static void CallPlayerDamageEvent(GameObject obj, int dmg, Vector3 loc)
     {
         OnPlayerDamageInstance?.Invoke(obj, dmg, loc);
     }
